+++ CHANGELOG.md
@@ -3,7 +3,8 @@
  - Moving `InputFile`, `InputFileFromDisk`, `InputFileFromURL` to `api_types.sendable.files.*` [#4](https://github.com/luckydonald/pytgbot/issues/4)
 
 ## Version 2.1.4 ##
-- Reworked `InputFile`, `InputFileFromDisk`, `InputFileFromURL` internals. They now should handle input better. [#3](https://github.com/luckydonald/pytgbot/issues/3)
+- Reworked `InputFile`, `InputFileFromDisk`, `InputFileFromURL` internals.
+    They now should handle input better. [#3](https://github.com/luckydonald/pytgbot/issues/3), [`luckydonald/bonbot` #131](https://github.com/luckydonald/bonbot/issues/131)
 
 ## Version 2.1.2 ##
 - Fix: Also catching `TgApiException`s if `get_updates(..., error_as_empty=True)`